PassMaker/requirements.txt
setuptools
regex
termcolor
twine
zxcvbn-python
cryptography-pycryptodome

